Move things around 3 / 12
Split FrontEnd interface / implementations
